Rename directory to just Image so we can nest multiple image Abilities
